# internal/webserver/swaggerdocs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for the honey web UI (`honey web`). Authenticate with the same token printed at startup, or `HONEY_WEB_TOKEN`. WebSocket endpoints `GET /ws/ssh` and `GET /ws/pve-qemu-vnc` are not described by this OpenAPI document.",
        "title": "Honey Web API",
        "contact": {
            "name": "honey"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8765",
    "basePath": "/",
    "paths": {
...
        "/api/v1/config/backends": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "List backends from config",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.Backends"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/config/backends/{kind}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Append backend entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "backend kind (gcp, aws, kubernetes, consul, proxmox, local)",
                        "name": "kind",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Backend entry JSON (shape depends on kind path param)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/webserver.ConfigBackendEntryBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/webserver.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/config/backends/{kind}/{index}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Replace backend entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "backend kind",
                        "name": "kind",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "0-based index in that kind's list",
                        "name": "index",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Backend entry JSON (shape depends on kind path param)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/webserver.ConfigBackendEntryBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/webserver.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Delete backend entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "backend kind",
                        "name": "kind",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "0-based index",
                        "name": "index",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/webserver.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/terminal-assist/models": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assist"
                ],
                "summary": "Terminal assist models",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/webserver.TerminalAssistModelsResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "config.AWSBackend": {
            "type": "object",
            "required": [
                "name",
                "profile"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "profile": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                }
            }
        },
...
        "config.Backends": {
            "type": "object",
            "properties": {
                "aws": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/config.AWSBackend"
                    }
                },
                "consul": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/config.ConsulBackend"
                    }
                },
                "gcp": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/config.GCPBackend"
                    }
                },
                "kubernetes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/config.KubernetesBackend"
                    }
                },
                "local": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/config.LocalBackend"
                    }
                },
                "proxmox": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/config.ProxmoxBackend"
                    }
                }
            }
        },
        "config.ConsulBackend": {
            "type": "object",
            "required": [
                "addr",
                "name"
            ],
            "properties": {
                "addr": {
                    "type": "string"
                },
                "datacenter": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "config.GCPBackend": {
            "type": "object",
            "required": [
                "name",
                "project"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "project": {
                    "type": "string"
                },
                "zone": {
                    "type": "string"
                }
            }
        },
        "config.KubernetesBackend": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "context": {
                    "type": "string"
                },
                "debug_image": {
                    "type": "string"
                },
                "kubeconfig": {
                    "type": "string"
                },
                "mode": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "config.LocalBackend": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "hosts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/config.LocalHost"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "config.LocalHost": {
            "type": "object",
            "required": [
                "name",
                "primary_ip"
            ],
            "properties": {
                "extra_ips": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "meta": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "primary_ip": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "zone": {
                    "type": "string"
                }
            }
        },
        "config.ProxmoxBackend": {
            "type": "object",
            "required": [
                "name",
                "url"
            ],
            "properties": {
                "exec_mode": {
                    "description": "ExecMode: ssh (default) = guest SSH for commands/SFTP/tunnels; pve = QEMU commands via guest agent API, LXC commands/SFTP over guest SSH (PVE has no LXC REST exec; web UI LXC console uses termproxy when token_id is set);\nhybrid = QEMU via guest agent + SSH for files; LXC uses guest SSH for commands and files.",
                    "type": "string"
                },
                "insecure": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "token_id": {
                    "type": "string"
                },
                "token_secret": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
...
        "webserver.ConfigBackendEntryBody": {
            "type": "object",
            "additionalProperties": true
        },
...
        "webserver.StatusResponse": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "webserver.TerminalAssistModelsResponse": {
            "type": "object",
            "properties": {
                "models": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Use `Authorization: Bearer \u003ctoken\u003e` where `\u003ctoken\u003e` is the web UI token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "HoneyTokenHeader": {
            "description": "Alternative header: `X-Honey-Token: \u003ctoken\u003e`.",
            "type": "apiKey",
            "name": "X-Honey-Token",
            "in": "header"
        },
        "TokenQuery": {
            "description": "Optional query token (same value) for URLs: `?token=\u003ctoken\u003e`.",
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    }
}